requirements.txt
requests
python-dotenv
feedparser
functions-framework==3.*
google-cloud
google-cloud-pubsub